+++ src/lib/types/HAPHostConfigType.ts
@@ -15,13 +15,39 @@ type HAPHostConfigType = NodeDef & {
     firmwareRev: SemVer
     hardwareRev: SemVer
     softwareRev: SemVer
+    bind?: string
+    bindType?: 'json' | 'str'
+    /**
+     * @deprecated use bind instead
+     */
     customMdnsConfig: boolean
+    /**
+     * @deprecated use bind instead
+     */
     mdnsMulticast: boolean
+    /**
+     * @deprecated use bind instead
+     */
     mdnsInterface: string
+    /**
+     * @deprecated use bind instead
+     */
     mdnsPort: number
+    /**
+     * @deprecated use bind instead
+     */
     mdnsIp: string
+    /**
+     * @deprecated use bind instead
+     */
     mdnsTtl: number
+    /**
+     * @deprecated use bind instead
+     */
     mdnsLoopback: boolean
+    /**
+     * @deprecated use bind instead
+     */
     mdnsReuseAddr: boolean
     allowMessagePassthrough: boolean
     accessoryCategory: HapCategories