+++ src/lib/types/hap-nodejs/BonjourMulticastOptions.ts
@@ -1,0 +1,11 @@
+type BonjourMulticastOptions = {
+    multicast?: boolean
+    interface?: string
+    port?: number
+    ip?: string
+    ttl?: number
+    loopback?: boolean
+    reuseAddr?: boolean
+}
+
+export default BonjourMulticastOptions